made all-products view visible
