+++ client/routes.js
@@ -2,7 +2,14 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import {Login, Signup, UserHome, AllUsers, Cart, AllProducts} from './components'
+import {
+  Login,
+  Signup,
+  UserHome,
+  AllUsers,
+  Cart,
+  AllProducts
+} from './components'
 import {me} from './store'
 
 /**
@@ -28,6 +35,7 @@ class Routes extends Component {
           <Switch>
             {/* Routes placed here are only available after logging in */}
             <Route path="/home" component={UserHome} />
+            <Route path="/products" component={AllProducts} />
           </Switch>
         )}
         {/* Displays our Login component as a fallback */}
